modified destroy

+++ jenkinsFiles/destory_kubernetes.groovy
@@ -10,13 +10,25 @@ node('slave1 || slave2') {
         """   
     }
 
-
-
-    stage("uninstall kondula app"){
+    stage("uninstall kandula service"){
         sh '''
-            echo "deleting service & app pods"
-            kubectl delete -f ./kubeFiles/kandula_service.yaml
-            kubectl delete -f ./kubeFiles/kandula_deploy.yaml
+            echo "deleting kandula service"
+            if kubectl get svc kandula-service; then
+                kubectl delete -f ./kubeFiles/kandula_service.yaml
+            else
+                echo "kandula service is not available"
+            fi
+        '''   
+    }
+
+    stage("uninstall kandula deployment") {
+        sh '''
+            echo "deleting kandula deployment"
+            if kubectl get deploy kandula-deployment; then
+                kubectl delete -f ./kubeFiles/kandula_deploy.yaml
+            else
+                echo "kandula deployment is not available"
+            fi
         '''   
     }
